requirements.txt
colorlog
colorama
requests
numpy
netCDF4
wrf-python